+++ install.sh
@@ -98,6 +98,8 @@ if [ -n "$PROFILE" ]; then
 else
 	info "Could not detect shell profile. Add these lines to your shell config:"
 	printf '\n  export PVM_DIR="%s"\n' "$PVM_DIR"
+
+	# shellcheck disable=SC2016
 	printf '  [ -s "$PVM_DIR/pvm.sh" ] && . "$PVM_DIR/pvm.sh"\n\n'
 fi